backend/foodgram/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Brotli==1.0.9
cffi==1.15.1
coreapi==2.3.3
cssselect2==0.7.0
Django==3.2
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.1.0
flake8==5.0.4
fonttools==4.38.0
html5lib==1.1
networkx==2.8.7
pdfkit==1.0.0
pep8==1.7.1
Pillow==9.2.0
pycodestyle==2.9.1
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.0
pytz==2022.4
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tinycss2==1.2.1
toml==0.10.2
webcolors==1.12
webencodings==0.5.1
zopfli==0.2.1
